3/7/12 - Mapped Skin to Profile instead of User - Added editSkin User interface (to get to it, go to account, and there is a link to the skin page where you can edit your skin's settings.) You can either write in the changes and click save, or press enter.
Note: some changes may not be immediately visable (for instance, changing
the parameters for buttons), you need to click the OpenBook link to go
back to the home page to see that change.

Note2: if the HTML does not use the main.css file, it will not get these
parametes (ex: it looks like the font size and type for links are set
elsewhere).

IMPORTANT!!!
Because these changes are not hardcoded into the initial-data.yml, if you
close your play run session, your changes will be lost!

Testing: added unit testing to test/unit/SkinTest.java. Not yet sure
about how to go about testing the user interface.